+++ src/graphql/queries.js
@@ -132,128 +132,3 @@ export const GET_PRIVATE_MESSAGES_BY_ID = gql`
     }
   }
 `;
-
-// ----------- MUTATIONS ----------------
-
-export const CREATE_PUBLIC_MESSAGE = gql`
-  mutation createPublicMessage($playerId: Int, $content: String) {
-    createPublicMessage(playerId: $playerId, content: $content) {
-      playerId {
-        id
-      }
-      content
-    }
-  }
-`;
-
-export const CLOSE_TRADE = gql`
-  mutation closeTrade($id: Int, $closed: Boolean) {
-    closeTrade(id: $id, closed: $closed) {
-      closed
-    }
-  }
-`;
-
-export const ACCEPT_TRADE = gql`
-  mutation acceptTrade(
-    $id: Int
-    $playerSenderId: Int
-    $playerReceiverId: Int
-    $moneyCashSender: Int
-    $moneyCashReceiver: Int
-    $eggSender: Int
-    $eggReceiver: Int
-    $featherSender: Int
-    $featherReceiver: Int
-    $bugSender: Int
-    $bugReceiver: Int
-  ) {
-    acceptTrade(
-      id: $id
-      playerSenderId: $playerSenderId
-      playerReceiverId: $playerReceiverId
-      moneyCashSender: $moneyCashSender
-      moneyCashReceiver: $moneyCashReceiver
-      eggSender: $eggSender
-      eggReceiver: $eggReceiver
-      featherSender: $featherSender
-      featherReceiver: $featherReceiver
-      bugSender: $bugSender
-      bugReceiver: $bugReceiver
-    ) {
-      id
-      playerSenderId {
-        id
-      }
-      playerReceiverId {
-        id
-      }
-      moneyCashSender
-      moneyCashReceiver
-      eggSender
-      eggReceiver
-      featherSender
-      featherReceiver
-      bugSender
-      bugReceiver
-    }
-  }
-`;
-
-export const SUGGEST_TRADE = gql`
-  mutation suggestTrade(
-    $playerSenderId: Int
-    $playerReceiverId: Int
-    $moneyCashSender: Int
-    $moneyCashReceiver: Int
-    $eggSender: Int
-    $eggReceiver: Int
-    $featherSender: Int
-    $featherReceiver: Int
-    $bugSender: Int
-    $bugReceiver: Int
-  ) {
-    suggestTrade(
-      playerSenderId: $playerSenderId
-      playerReceiverId: $playerReceiverId
-      moneyCashSender: $moneyCashSender
-      moneyCashReceiver: $moneyCashReceiver
-      eggSender: $eggSender
-      eggReceiver: $eggReceiver
-      featherSender: $featherSender
-      featherReceiver: $featherReceiver
-      bugSender: $bugSender
-      bugReceiver: $bugReceiver
-    ) {
-      playerSenderId {
-        id
-      }
-      playerReceiverId {
-        id
-      }
-      moneyCashSender
-      moneyCashReceiver
-      eggSender
-      eggReceiver
-      featherSender
-      featherReceiver
-      bugSender
-      bugReceiver
-    }
-  }
-`;
-
-// ----------- SUBSCRIPTIONS ------------
-
-export const SUB_ALL_PUBLIC_MESSAGES = gql`
-  subscription getAllPublicMessages {
-    getAllPublicMessages {
-      id
-      content
-      playerId {
-        id
-        name
-      }
-    }
-  }
-`;
