+++ src/graphql/subscriptions.js
@@ -1,0 +1,12 @@
+export const SUB_ALL_PUBLIC_MESSAGES = gql`
+  subscription getAllPublicMessages {
+    getAllPublicMessages {
+      id
+      content
+      playerId {
+        id
+        name
+      }
+    }
+  }
+`;
