+++ src/index.js
@@ -18,19 +18,15 @@ import { apiUrl } from "./config/constants";
 import { onError } from "apollo-link-error";
 import { ApolloLink } from "apollo-link";
 
-// const errorLink = onError(({ graphQLErrors, networkError }) => {
-//   if (graphQLErrors) {
-//     console.log("GRAPHQL ERROR:", graphQLErrors);
-//   }
-//   if (networkError) {
-//     console.log("NETWORKERROR", networkError);
-//   }
-// });
-
 const httpLink = new HttpLink({
   uri: `http://${apiUrl}`,
-  // credentials: "include",
+  // credentials: "same-origin",
 });
+
+// const httpLink = new HttpLink({
+// uri: `http://${apiUrl}`,
+// credentials: "include",
+// });
 
 // Create a WebSocket link:
 const wsLink = new WebSocketLink({
@@ -39,8 +35,6 @@ const wsLink = new WebSocketLink({
     reconnect: true,
   },
 });
-
-// const link2 = ApolloLink.from([errorLink, httpLink]);
 
 const link = split(
   ({ query }) => {
@@ -54,9 +48,8 @@ const link = split(
 
 // Instantiate client
 const client = new ApolloClient({
-  // link2,
-  link,
   cache: new InMemoryCache(),
+  link,
 });
 
 ReactDOM.render(
